+++ examples/example-utils.js
@@ -1,7 +1,9 @@
+const craftx = require('../src');
+
 const snooze = ms => new Promise(resolve => setTimeout(resolve, ms));
 
 const queryUserInfo = async () => {
-  await snooze(800);
+  //await snooze(800);
   return {
     name: 'John',
     dob: '25/03/1986',
@@ -9,7 +11,7 @@ const queryUserInfo = async () => {
 };
 
 const queryAccountInfo = async (user) => {
-  await snooze(500);
+  //await snooze(500);
   return {
     preferences: 32,
     permissions: 7,
@@ -17,13 +19,21 @@ const queryAccountInfo = async (user) => {
   };
 };
 
-const thisFnThrows = async () => {
-  throw new Error('thisFnThrows Error');
+const thisFnThrows = async (doThrow = true) => {
+  if (doThrow) {
+    throw new Error('thisFnThrows Error');
+  }
+
+  return 123;
+};
+
+const simple = () => {
+  return 'abc';
 };
 
 // example of independant
 const querySystemInfo = async () => {
-  await snooze(200);
+  //await snooze(200);
   return {
     load: 14.0,
   };
@@ -31,7 +41,7 @@ const querySystemInfo = async () => {
 
 // example of promise
 const queryConfig = (sender) => new Promise(async (resolve) => {
-  await snooze(300);
+  //await snooze(300);
   resolve({
     url: 'https://amrayn.com',
     sender,
@@ -44,4 +54,5 @@ module.exports = {
   queryAccountInfo,
   thisFnThrows,
   queryConfig,
+  simple,
 };