+++ CHANGELOG.md
@@ -1,4 +1,8 @@
-## 2.0.0
+## 2.0.2
+- Changed default export to `fn()`
+- Added `fnExport` helper function
+
+## 2.0.1
 - Migration from makefun
 - Resolution for typed arrays and other built-in data structure types
 - More formal test for values